Add Different Speaker in Shared Info topic
